+++ R/bflSmooth.R
@@ -18,10 +18,7 @@ weights_control <- function(weights,start,hf_length,hf_freq) {
 }
 
 bflSmooth_matrices_impl <- function(lf_length,ratio,weights,lfserie.is.rate) {
-  if (is.null(weights)) {
-    aggregated_weights <- rep(1,lf_length)
-    weights <- 1
-  }
+  if (is.null(weights)) weights <- 1
   else {
     aggregated_weights <- aggregate.ts(weights,frequency(weights)/ratio)
     weights <- weights/ts_expand(aggregated_weights,
@@ -38,7 +35,7 @@ bflSmooth_matrices_impl <- function(lf_length,ratio,weights,lfserie.is.rate) {
   # A * as.numeric(x) stands for diag(x) %*% A
   # t(t(A) * as.numeric(x)) stands for A %*% diag(x)
   
-  if (lfserie.is.rate||is.null(weights)) {
+  if (lfserie.is.rate||identical(weights,1)) {
     list(m1=m1,
          cprod1=cprod1,
          cprod2=cprod2)
@@ -50,7 +47,7 @@ bflSmooth_matrices_impl <- function(lf_length,ratio,weights,lfserie.is.rate) {
 }
 
 # This function generates a wrapper of bflSmooth_matrices_impl that gives the
-# same results but uses cache, which is useful considering a lot of similar
+# same results but uses cache, which is useful considering a lot of similar calls
 bflSmooth_matrices_generator <- function(cache_size=100L) {
   cache <- vector("list",cache_size)
   cache_next <- 1L
@@ -79,11 +76,9 @@ bflSmooth_matrices <- bflSmooth_matrices_generator()
 #' If `weights` isn't `NULL` the results depends of `lfserie.is.rate` :
 #' 
 #' * if `FALSE` the rate output/weights is smoothed with the constraint that the aggregated output
-#' is equal to lfserie.
-#' * if `TRUE` the output is the rate to be smoothed, and the output is the aggregated rate.
-#' 
-#' @author
-#' Arnaud Feldmann
+#' is equal to the input lfserie.
+#' * if `TRUE` the input lfserie is the rate to be smoothed, with the constraint that it respects
+#' the products lfserie*weights through aggregation.
 #' 
 #' @param lfserie a time-serie to be smoothed
 #' @param nfrequency the new high frequency. It must be a multiple of the low frequency.
@@ -102,6 +97,7 @@ bflSmooth <- function(lfserie,nfrequency,weights=NULL,lfserie.is.rate=FALSE) {
   if (!is.null(dim(lfserie)) && dim(lfserie)[2L] != 1) stop("The low frequency serie must be one-dimensional", call. = FALSE)
   if (is.null(weights) && lfserie.is.rate) {
     warning("weights is NULL. Ignoring lfserie.is.rate")
+    lfserie.is.rate <- FALSE
   }
   
   ratio <- nfrequency/tsplf[3L]